Extract random-local-path helper
Extracting this function to reuse it for unit tests as well.
